rename: newCardPage 폴더 생성
