+++ src/components/newCardPage/newCardInputSection/NewCardInputSection.styled.ts
