+++ src/components/newCardPage/newCardInputSection/NewCardInputSection.tsx
